Create README.md
Welcome to Tigray Home Town, a repository dedicated to showcasing the culture, history and heritage of Tigray, a region in norther Ethiopia known for its rich traditions, breathtaking landscapes, and deep historical significance. 
This project aims to serve as virtual space for sharing knowledge, fostering community, and celebrating the beauty of Tigray and its people.